Refactor StreamValidator out of core
