+++ core/src/test/resources/config-dev.yaml
@@ -1,7 +1,7 @@
 env: test
 
 streamValidatorConfig:
-  className: com.homeaway.streamingplatform.extensions.validator.StreamValidatorIT$ValidProductStreamValidator
+  className: com.homeaway.streamingplatform.resource.StreamResourceValidationIT$ValidProductStreamValidator
   properties:
     validator.invalid-id: 13
