+++ validator-streams/src/main/java/com/homeaway/streamingplatform/validator/streams/EmptyStreamValidator.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.homeaway.streamingplatform.extensions.validation;
+package com.homeaway.streamingplatform.validator.streams;
 
 import java.util.Map;
